feat: Add DiagramBuilder for parsing and rendering diagrams
This commit introduces a new `DiagramBuilder` class to handle the parsing and rendering of PPTX XML diagrams. The `DiagramBuilder` is responsible for parsing both drawing-based and layout-based (SmartArt) diagrams.

The new class follows a similar structure to the existing `ShapeBuilder` and encapsulates the diagram parsing logic. The `slideHandler` has been refactored to use the `DiagramBuilder`, and the old `diagramParser.js` file has been removed.

The implementation includes support for the linear layout algorithm and a basic implementation of the spiral layout algorithm. It also handles the `ch` (child) axis for traversing the data model.
