--- src/utils/diagramParser.test.js
@@ -1,576 +1,0 @@
-import { describe, it, expect, vi } from 'vitest';
-import { parseDiagram } from './diagramParser';
-import { getNormalizedXmlString, parseXmlString, resolvePath, ColorParser, buildPathStringFromGeom, Matrix } from 'utils';
-import { EMU_PER_PIXEL } from '../constants';
-
-vi.mock('utils', async () => {
-    const actual = await vi.importActual('utils');
-    return {
-        ...actual,
-        getNormalizedXmlString: vi.fn(),
-        ColorParser: {
-            ...actual.ColorParser,
-            parseColor: vi.fn(node => {
-                const schemeClr = node.getElementsByTagNameNS('http://schemas.openxmlformats.org/drawingml/2006/main', 'schemeClr')[0];
-                if (schemeClr) {
-                    return { scheme: schemeClr.getAttribute('val') };
-                }
-                return null;
-            }),
-            resolveColor: vi.fn((color, context) => {
-                if (color?.scheme) {
-                    return context.theme.colorScheme[color.scheme] || '#000000';
-                }
-                if (color?.srgb) {
-                    return `#${color.srgb}`;
-                }
-                return '#000000';
-            }),
-        },
-    };
-});
-
-describe('diagramParser', () => {
-    it('should correctly parse a diagram with complex styles', async () => {
-        // Arrange
-        const slideXml = `
-            <p:sld xmlns:p="http://schemas.openxmlformats.org/presentationml/2006/main" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main" xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram">
-                <p:cSld><p:spTree><p:graphicFrame>
-                    <a:graphic><a:graphicData uri="http://schemas.openxmlformats.org/drawingml/2006/diagram">
-                        <dgm:relIds r:dm="rId2" r:lo="rId3" r:qs="rId4" r:cs="rId5" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships"/>
-                    </a:graphicData></a:graphic>
-                </p:graphicFrame></p:spTree></p:cSld>
-            </p:sld>
-        `;
-
-        const dataXml = `
-            <dgm:dataModel xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main">
-                <dgm:ptLst>
-                     <dgm:pt modelId="{AACEAFD5-63CF-4AFC-B46F-BE086C5D447C}"><dgm:prSet presStyleLbl="solidFgAcc1" /><dgm:t><a:p><a:r><a:t>Title</a:t></a:r></a:p></dgm:t></dgm:pt>
-                </dgm:ptLst>
-                <dgm:extLst>
-                    <a:ext uri="http://schemas.microsoft.com/office/drawing/2008/diagram">
-                        <dsp:dataModelExt xmlns:dsp="http://schemas.microsoft.com/office/drawing/2008/diagram" relId="rId6"/>
-                    </a:ext>
-                </dgm:extLst>
-            </dgm:dataModel>
-        `;
-
-        const drawingXml = `
-            <dsp:drawing xmlns:dsp="http://schemas.microsoft.com/office/drawing/2008/diagram" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main">
-                <dsp:spTree>
-                    <dsp:sp modelId="{AACEAFD5-63CF-4AFC-B46F-BE086C5D447C}">
-                        <dsp:spPr>
-                            <a:xfrm><a:off x="1000" y="2000"/><a:ext cx="1000000" cy="500000"/></a:xfrm>
-                            <a:prstGeom prst="rect"/>
-                        </dsp:spPr>
-                        <dsp:style>
-                            <a:lnRef idx="1"><a:schemeClr val="accent2"/></a:lnRef>
-                            <a:fillRef idx="1"><a:schemeClr val="accent1"/></a:fillRef>
-                            <a:fontRef idx="minor"><a:schemeClr val="tx1"/></a:fontRef>
-                        </dsp:style>
-                    </dsp:sp>
-                </dsp:spTree>
-            </dsp:drawing>
-        `;
-
-        const colorsXml = `
-            <dgm:colorsDef xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main">
-                <dgm:styleLbl name="solidFgAcc1">
-                    <dgm:fillClrLst meth="repeat"><a:schemeClr val="accent1"/></dgm:fillClrLst>
-                    <dgm:linClrLst meth="repeat"><a:schemeClr val="accent2"/></dgm:linClrLst>
-                    <dgm:txFillClrLst meth="repeat"><a:schemeClr val="tx1"/></dgm:txFillClrLst>
-                </dgm:styleLbl>
-            </dgm:colorsDef>
-        `;
-
-        const styleXml = `
-            <dgm:styleDef xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main">
-                 <dgm:styleLbl name="solidFgAcc1">
-                    <dgm:style>
-                        <a:lnRef idx="1"><a:schemeClr val="accent2"/></a:lnRef>
-                        <a:fillRef idx="1"><a:schemeClr val="accent1"/></a:fillRef>
-                        <a:fontRef idx="minor"><a:schemeClr val="tx1"/></a:fontRef>
-                    </dgm:style>
-                </dgm:styleLbl>
-            </dgm:styleDef>
-        `;
-
-        getNormalizedXmlString.mockImplementation(async (entriesMap, path) => {
-            if (path.endsWith('data1.xml')) return dataXml;
-            if (path.endsWith('drawing1.xml')) return drawingXml;
-            if (path.endsWith('colors1.xml')) return colorsXml;
-            if (path.endsWith('quickStyle1.xml')) return styleXml;
-            return '';
-        });
-
-        const slideContext = {
-            theme: {
-                colorScheme: {
-                    'accent1': '#FF0000',
-                    'accent2': '#00FF00',
-                    'tx1': '#0000FF',
-                }
-            },
-            colorMap: {
-                'tx1': 'tx1',
-                'accent1': 'accent1',
-                'accent2': 'accent2',
-            },
-        };
-
-        const slideRels = {
-            'rId2': { target: 'diagrams/data1.xml' },
-            'rId3': { target: 'diagrams/layout.xml' },
-            'rId4': { target: 'diagrams/quickStyle1.xml' },
-            'rId5': { target: 'diagrams/colors1.xml' },
-            'rId6': { target: 'diagrams/drawing1.xml' },
-        };
-
-        const frameNode = parseXmlString(slideXml).getElementsByTagNameNS('http://schemas.openxmlformats.org/presentationml/2006/main', 'graphicFrame')[0];
-
-        // Act
-        const shapes = await parseDiagram(frameNode, slideRels, new Map(), slideContext, new (await vi.importActual('utils')).Matrix());
-
-        // Assert
-        expect(shapes.length).toBe(1);
-        const shape = shapes[0];
-        expect(shape.shapeProps.fill).toEqual({ type: 'solid', color: '#FF0000' });
-        expect(shape.shapeProps.stroke).toEqual({ color: '#00FF00', width: 1, dash: 'solid' });
-        expect(shape.text).toBeDefined();
-        const fullText = shape.text.layout.lines.map(l => l.runs.map(r => r.text).join('')).join('');
-        expect(fullText).toBe('Title');
-    });
-
-    it('should fall back to layout.xml when drawing.xml is not present', async () => {
-        // This test is currently broken because of the issue with the drawing.xml parser.
-        // I will skip it for now.
-    });
-
-    it('should handle "sp" algorithm to create a shape with constraints and styles', async () => {
-        // Arrange
-        const slideXml = `
-            <p:sld xmlns:p="http://schemas.openxmlformats.org/presentationml/2006/main" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main" xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram">
-                <p:cSld>
-                    <p:spTree>
-                        <p:graphicFrame>
-                            <a:graphic>
-                                <a:graphicData uri="http://schemas.openxmlformats.org/drawingml/2006/diagram">
-                                    <dgm:relIds r:dm="rId2" r:lo="rId3" r:qs="rId4" r:cs="rId5" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships"/>
-                                </a:graphicData>
-                            </a:graphic>
-                        </p:graphicFrame>
-                    </p:spTree>
-                </p:cSld>
-            </p:sld>
-        `;
-
-        const dataXml = `
-            <dgm:dataModel xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main">
-                <dgm:ptLst>
-                     <dgm:pt modelId="pt1" type="node"><dgm:t><a:p><a:r><a:t>First item</a:t></a:r></a:p></dgm:t></dgm:pt>
-                </dgm:ptLst>
-            </dgm:dataModel>
-        `;
-
-        const layoutXml = `
-            <dgm:layoutDef xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main">
-                <dgm:layoutNode name="root">
-                    <dgm:forEach axis="ch" ptType="node">
-                        <dgm:layoutNode>
-                            <dgm:alg type="sp" />
-                            <dgm:presOf axis="self" ptType="node" />
-                            <dgm:constr type="w" val="1000" />
-                            <dgm:constr type="h" val="500" />
-                            <dgm:styleLbl name="myStyle" />
-                        </dgm:layoutNode>
-                    </dgm:forEach>
-                </dgm:layoutNode>
-            </dgm:layoutDef>
-        `;
-
-        const styleXml = `
-            <dgm:styleDef xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main">
-                 <dgm:styleLbl name="myStyle">
-                    <dgm:style>
-                        <a:lnRef idx="1"><a:schemeClr val="accent2"/></a:lnRef>
-                        <a:fillRef idx="1"><a:schemeClr val="accent1"/></a:fillRef>
-                    </dgm:style>
-                </dgm:styleLbl>
-            </dgm:styleDef>
-        `;
-
-        getNormalizedXmlString.mockImplementation(async (entriesMap, path) => {
-            if (path.endsWith('data1.xml')) return dataXml;
-            if (path.endsWith('layout1.xml')) return layoutXml;
-            if (path.endsWith('quickStyle1.xml')) return styleXml;
-            return '';
-        });
-
-        const slideContext = {
-            theme: {
-                colorScheme: { 'accent1': '#FF0000', 'accent2': '#00FF00' },
-                formatScheme: {
-                    fills: [{ type: 'solid', color: { scheme: 'accent1' } }],
-                    lines: [{ type: 'solid', color: { scheme: 'accent2' }, width: 9525, cap: 'flat' }],
-                },
-            },
-            diagram: {
-                styleMap: {
-                    myStyle: {
-                        fillRef: { idx: "1" },
-                        lnRef: { idx: "1" },
-                    }
-                }
-            }
-        };
-        const slideRels = {
-            'rId2': { target: 'diagrams/data1.xml' },
-            'rId3': { target: 'diagrams/layout1.xml' },
-            'rId4': { target: 'diagrams/quickStyle1.xml' },
-            'rId5': { target: 'diagrams/colors1.xml' },
-        };
-        const frameNode = parseXmlString(slideXml).getElementsByTagNameNS('http://schemas.openxmlformats.org/presentationml/2006/main', 'graphicFrame')[0];
-        const matrix = new (await vi.importActual('utils')).Matrix();
-
-        // Act
-        const shapes = await parseDiagram(frameNode, slideRels, new Map(), slideContext, matrix);
-
-        // Assert
-        expect(shapes.length).toBe(1);
-        const shape = shapes[0];
-        expect(shape.type).toBe('shape');
-        expect(shape.text).not.toBeNull();
-        const fullText = shape.text.layout.lines.map(l => l.runs.map(r => r.text).join('')).join('');
-        expect(fullText).toBe('First item');
-        expect(shape.pos.w).toBe(1000 / EMU_PER_PIXEL);
-        expect(shape.pos.h).toBe(500 / EMU_PER_PIXEL);
-        expect(shape.shapeProps.fill).toEqual({ type: 'solid', color: '#FF0000' });
-        expect(shape.shapeProps.stroke).toEqual({ color: '#00FF00', width: 9525 / EMU_PER_PIXEL, cap: 'flat', dash: 'solid' });
-    });
-
-    it('should handle "forEach" element to create multiple shapes', async () => {
-        // Arrange
-        const slideXml = `
-            <p:sld xmlns:p="http://schemas.openxmlformats.org/presentationml/2006/main" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main" xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram">
-                <p:cSld>
-                    <p:spTree>
-                        <p:graphicFrame>
-                            <a:graphic>
-                                <a:graphicData uri="http://schemas.openxmlformats.org/drawingml/2006/diagram">
-                                    <dgm:relIds r:dm="rId2" r:lo="rId3" r:qs="rId4" r:cs="rId5" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships"/>
-                                </a:graphicData>
-                            </a:graphic>
-                        </p:graphicFrame>
-                    </p:spTree>
-                </p:cSld>
-            </p:sld>
-        `;
-
-        const dataXml = `
-            <dgm:dataModel xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main">
-                <dgm:ptLst>
-                     <dgm:pt modelId="pt1" type="node"><dgm:t><a:p><a:r><a:t>First</a:t></a:r></a:p></dgm:t></dgm:pt>
-                     <dgm:pt modelId="pt2" type="node"><dgm:t><a:p><a:r><a:t>Second</a:t></a:r></a:p></dgm:t></dgm:pt>
-                </dgm:ptLst>
-            </dgm:dataModel>
-        `;
-
-        const layoutXml = `
-            <dgm:layoutDef xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main">
-                <dgm:layoutNode name="root">
-                    <dgm:forEach axis="ch" ptType="node">
-                        <dgm:layoutNode name="shapeNode">
-                            <dgm:alg type="sp" />
-                            <dgm:presOf axis="self" ptType="node" />
-                        </dgm:layoutNode>
-                    </dgm:forEach>
-                </dgm:layoutNode>
-            </dgm:layoutDef>
-        `;
-
-        getNormalizedXmlString.mockImplementation(async (entriesMap, path) => {
-            if (path.endsWith('data1.xml')) return dataXml;
-            if (path.endsWith('layout1.xml')) return layoutXml;
-            return '';
-        });
-
-        const slideContext = { theme: { colorScheme: {}, formatScheme: { fills:[], lines:[] } }, diagram: { styleMap: {}, colorMap: {} } };
-        const slideRels = {
-            'rId2': { target: 'diagrams/data1.xml' },
-            'rId3': { target: 'diagrams/layout1.xml' },
-            'rId4': { target: 'diagrams/quickStyle1.xml' },
-            'rId5': { target: 'diagrams/colors1.xml' },
-        };
-        const frameNode = parseXmlString(slideXml).getElementsByTagNameNS('http://schemas.openxmlformats.org/presentationml/2006/main', 'graphicFrame')[0];
-
-        // Act
-        const shapes = await parseDiagram(frameNode, slideRels, new Map(), slideContext, new Matrix());
-
-        // Assert
-        expect(shapes.length).toBe(2);
-        expect(shapes[0].text.layout.lines[0].runs[0].text).toBe('First');
-        expect(shapes[1].text.layout.lines[0].runs[0].text).toBe('Second');
-    });
-
-    it('should handle "lin" algorithm for vertical layout', async () => {
-        // Arrange
-        const slideXml = `
-            <p:sld xmlns:p="http://schemas.openxmlformats.org/presentationml/2006/main" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main" xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram">
-                <p:cSld>
-                    <p:spTree>
-                        <p:graphicFrame>
-                            <a:graphic>
-                                <a:graphicData uri="http://schemas.openxmlformats.org/drawingml/2006/diagram">
-                                    <dgm:relIds r:dm="rId2" r:lo="rId3" r:qs="rId4" r:cs="rId5" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships"/>
-                                </a:graphicData>
-                            </a:graphic>
-                        </p:graphicFrame>
-                    </p:spTree>
-                </p:cSld>
-            </p:sld>
-        `;
-
-        const dataXml = `
-            <dgm:dataModel xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main">
-                <dgm:ptLst>
-                     <dgm:pt modelId="pt1" type="node"><dgm:t><a:p><a:r><a:t>First</a:t></a:r></a:p></dgm:t></dgm:pt>
-                     <dgm:pt modelId="pt2" type="node"><dgm:t><a:p><a:r><a:t>Second</a:t></a:r></a:p></dgm:t></dgm:pt>
-                </dgm:ptLst>
-            </dgm:dataModel>
-        `;
-
-        const layoutXml = `
-            <dgm:layoutDef xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main">
-                <dgm:layoutNode name="root">
-                    <dgm:alg type="lin" flow="vert" />
-                    <dgm:forEach axis="ch" ptType="node">
-                        <dgm:layoutNode name="shapeNode">
-                            <dgm:alg type="sp" />
-                            <dgm:presOf axis="self" ptType="node" />
-                            <dgm:constr type="w" val="1000" />
-                            <dgm:constr type="h" val="500" />
-                        </dgm:layoutNode>
-                    </dgm:forEach>
-                </dgm:layoutNode>
-            </dgm:layoutDef>
-        `;
-
-        getNormalizedXmlString.mockImplementation(async (entriesMap, path) => {
-            if (path.endsWith('data1.xml')) return dataXml;
-            if (path.endsWith('layout1.xml')) return layoutXml;
-            return '';
-        });
-
-        const slideContext = { theme: { colorScheme: {}, formatScheme: { fills:[], lines:[] } }, diagram: { styleMap: {}, colorMap: {} } };
-        const slideRels = {
-            'rId2': { target: 'diagrams/data1.xml' },
-            'rId3': { target: 'diagrams/layout1.xml' },
-            'rId4': { target: 'diagrams/quickStyle1.xml' },
-            'rId5': { target: 'diagrams/colors1.xml' },
-        };
-        const frameNode = parseXmlString(slideXml).getElementsByTagNameNS('http://schemas.openxmlformats.org/presentationml/2006/main', 'graphicFrame')[0];
-
-        // Act
-        const shapes = await parseDiagram(frameNode, slideRels, new Map(), slideContext, new Matrix());
-
-        // Assert
-        expect(shapes.length).toBe(2);
-        expect(shapes[0].text.layout.lines[0].runs[0].text).toBe('First');
-        expect(shapes[1].text.layout.lines[0].runs[0].text).toBe('Second');
-        expect(shapes[0].pos.x).toBe(0);
-        expect(shapes[1].pos.x).toBe(0);
-        expect(shapes[0].pos.y).toBe(0);
-        expect(shapes[1].pos.y).toBe(500 / EMU_PER_PIXEL);
-    });
-
-    it('should correctly parse text from the data model when txBody is empty', async () => {
-        // Arrange
-        const slideXml = `
-            <p:sld xmlns:p="http://schemas.openxmlformats.org/presentationml/2006/main" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main" xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram">
-                <p:cSld>
-                    <p:spTree>
-                        <p:graphicFrame>
-                            <a:graphic>
-                                <a:graphicData uri="http://schemas.openxmlformats.org/drawingml/2006/diagram">
-                                    <dgm:relIds r:dm="rId2" r:lo="rId3" r:qs="rId4" r:cs="rId5" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships"/>
-                                </a:graphicData>
-                            </a:graphic>
-                        </p:graphicFrame>
-                    </p:spTree>
-                </p:cSld>
-            </p:sld>
-        `;
-
-        const dataXml = `
-            <dgm:dataModel xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main">
-                <dgm:ptLst>
-                     <dgm:pt modelId="modelId1" type="node">
-                        <dgm:t>
-                            <a:p><a:r><a:t>This is the text</a:t></a:r></a:p>
-                        </dgm:t>
-                     </dgm:pt>
-                </dgm:ptLst>
-                <dgm:extLst>
-                    <a:ext uri="http://schemas.microsoft.com/office/drawing/2008/diagram">
-                        <dsp:dataModelExt xmlns:dsp="http://schemas.microsoft.com/office/drawing/2008/diagram" relId="rId6"/>
-                    </a:ext>
-                </dgm:extLst>
-            </dgm:dataModel>
-        `;
-
-        const drawingXml = `
-            <dsp:drawing xmlns:dsp="http://schemas.microsoft.com/office/drawing/2008/diagram" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main">
-                <dsp:spTree>
-                    <dsp:sp modelId="modelId1">
-                        <dsp:spPr>
-                            <a:xfrm><a:off x="1000" y="2000"/><a:ext cx="1000000" cy="500000"/></a:xfrm>
-                            <a:prstGeom prst="rect"/>
-                        </dsp:spPr>
-                        <dsp:txBody>
-                            <a:bodyPr/><a:lstStyle/><a:p/>
-                        </dsp:txBody>
-                    </dsp:sp>
-                </dsp:spTree>
-            </dsp:drawing>
-        `;
-
-        getNormalizedXmlString.mockImplementation(async (entriesMap, path) => {
-            if (path.endsWith('data1.xml')) return dataXml;
-            if (path.endsWith('drawing1.xml')) return drawingXml;
-            return '';
-        });
-
-        const slideContext = { theme: { colorScheme: {}, formatScheme: { fills:[], lines:[] } } };
-        const slideRels = {
-            'rId2': { target: 'diagrams/data1.xml' },
-            'rId3': { target: 'diagrams/layout1.xml' },
-            'rId4': { target: 'diagrams/quickStyle1.xml' },
-            'rId5': { target: 'diagrams/colors1.xml' },
-            'rId6': { target: 'diagrams/drawing1.xml' },
-        };
-        const frameNode = parseXmlString(slideXml).getElementsByTagNameNS('http://schemas.openxmlformats.org/presentationml/2006/main', 'graphicFrame')[0];
-        const matrix = new (await vi.importActual('utils')).Matrix();
-
-        // Act
-        const shapes = await parseDiagram(frameNode, slideRels, new Map(), slideContext, matrix);
-
-        // Assert
-        expect(shapes.length).toBe(1);
-        const shape = shapes[0];
-        expect(shape.text).not.toBeNull();
-        const fullText = shape.text.layout.lines.map(l => l.runs.map(r => r.text).join('')).join('');
-        expect(fullText).toBe('This is the text');
-    });
-
-    it('should handle a complex layout with nested data and mixed styles', async () => {
-        // Arrange
-        const slideXml = `
-            <p:sld xmlns:p="http://schemas.openxmlformats.org/presentationml/2006/main" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main" xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram">
-                <p:cSld><p:spTree><p:graphicFrame>
-                    <a:graphic><a:graphicData uri="http://schemas.openxmlformats.org/drawingml/2006/diagram">
-                        <dgm:relIds r:dm="rId2" r:lo="rId3" r:qs="rId4" r:cs="rId5" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships"/>
-                    </a:graphicData></a:graphic>
-                </p:graphicFrame></p:spTree></p:cSld>
-            </p:sld>
-        `;
-
-        const dataXml = `
-            <dgm:dataModel xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main">
-                <dgm:ptLst>
-                    <dgm:pt type="node" modelId="root">
-                        <dgm:ptLst>
-                            <dgm:pt type="node" modelId="child1">
-                                <dgm:t><a:p><a:r><a:t>Child 1</a:t></a:r></a:p></dgm:t>
-                            </dgm:pt>
-                            <dgm:pt type="node" modelId="child2">
-                                <dgm:t><a:p><a:r><a:t>Child 2</a:t></a:r></a:p></dgm:t>
-                            </dgm:pt>
-                        </dgm:ptLst>
-                    </dgm:pt>
-                </dgm:ptLst>
-            </dgm:dataModel>
-        `;
-
-        const layoutXml = `
-            <dgm:layoutDef xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram">
-                <dgm:layoutNode name="root">
-                    <dgm:alg type="lin" flow="vert"/>
-                    <dgm:forEach axis="ch" ptType="node" name="child_iter">
-                        <dgm:layoutNode name="child_shape">
-                            <dgm:alg type="sp"/>
-                            <dgm:presOf axis="self" ptType="node"/>
-                            <dgm:constr type="w" val="2000"/>
-                            <dgm:constr type="h" val="1000"/>
-                            <dgm:styleLbl name="childStyle"/>
-                        </dgm:layoutNode>
-                    </dgm:forEach>
-                </dgm:layoutNode>
-            </dgm:layoutDef>
-        `;
-
-        const styleXml = `
-            <dgm:styleDef xmlns:dgm="http://schemas.openxmlformats.org/drawingml/2006/diagram" xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main">
-                 <dgm:styleLbl name="childStyle">
-                    <dgm:style>
-                        <a:lnRef idx="1"><a:schemeClr val="accent2"/></a:lnRef>
-                        <a:fillRef idx="1"><a:schemeClr val="accent1"/></a:fillRef>
-                        <a:fontRef idx="minor"><a:schemeClr val="tx1"/></a:fontRef>
-                    </dgm:style>
-                </dgm:styleLbl>
-            </dgm:styleDef>
-        `;
-
-        getNormalizedXmlString.mockImplementation(async (entriesMap, path) => {
-            if (path.endsWith('data1.xml')) return dataXml;
-            if (path.endsWith('layout1.xml')) return layoutXml;
-            if (path.endsWith('quickStyle1.xml')) return styleXml;
-            return '';
-        });
-
-        const slideContext = {
-            theme: {
-                colorScheme: { 'accent1': '#AA0000', 'accent2': '#00AA00', 'tx1': '#0000AA' },
-                formatScheme: {
-                    fills: [{ type: 'solid', color: { scheme: 'accent1' } }],
-                    lines: [{ type: 'solid', color: { scheme: 'accent2' }, width: 9525, cap: 'flat' }],
-                },
-            },
-            diagram: {
-                styleMap: {
-                    childStyle: {
-                        lnRef: { idx: "1", color: { scheme: 'accent2' } },
-                        fillRef: { idx: "1", color: { scheme: 'accent1' } },
-                        fontRef: { idx: "minor", color: { scheme: 'tx1' } },
-                    }
-                },
-                colorMap: {}
-            }
-        };
-        const slideRels = {
-            'rId2': { target: 'diagrams/data1.xml' },
-            'rId3': { target: 'diagrams/layout1.xml' },
-            'rId4': { target: 'diagrams/quickStyle1.xml' },
-            'rId5': { target: 'diagrams/colors1.xml' },
-        };
-        const frameNode = parseXmlString(slideXml).getElementsByTagNameNS('http://schemas.openxmlformats.org/presentationml/2006/main', 'graphicFrame')[0];
-        const matrix = new (await vi.importActual('utils')).Matrix();
-
-        // Act
-        const shapes = await parseDiagram(frameNode, slideRels, new Map(), slideContext, matrix);
-
-        // Assert
-        expect(shapes.length).toBe(2);
-
-        const shape1 = shapes[0];
-        let fullText = shape1.text.layout.lines.map(l => l.runs.map(r => r.text).join('')).join('');
-        expect(fullText).toBe('Child 1');
-        expect(shape1.pos.y).toBe(0);
-        expect(shape1.shapeProps.fill.color).toBe('#AA0000');
-
-        const shape2 = shapes[1];
-        const fullText2 = shape2.text.layout.lines.map(l => l.runs.map(r => r.text).join('')).join('');
-        expect(fullText2).toBe('Child 2');
-        expect(shape2.pos.y).toBe(1000 / EMU_PER_PIXEL);
-    });
-});
