+++ src/utils/index.js
@@ -12,6 +12,6 @@ export { getRelationships } from './getRelationships.js';
 export { getSlideOrder } from './getSlideOrder.js';
 export { getSlideSize } from './getSlideSize.js';
 export { parseExtensions } from './extensionParser.js';
-export { parseDiagram } from './diagramParser.js';
+export { DiagramBuilder } from './diagramBuilder.js';
 export * from './pptxParser.js';
 export * from './drawing.js';
